requirements.txt
aiogram==2.22.2
python-dotenv==0.21.0
gino==1.0.1
asyncpg==0.26.0
apscheduler==3.9.1
et-xmlfile==1.1.0
numpy==1.23.3
sqlalchemy==1.4.41
httplib2==0.20.4